+++ apps/dashboard/views/playlists.py
@@ -1,9 +1,13 @@
+from copy import copy
+
+from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.core.exceptions import ValidationError, PermissionDenied
 from django.shortcuts import render, redirect, get_object_or_404
+from django.urls import reverse
 from django_tables2 import RequestConfig
 
-from apps.dashboard.forms import DashboardAddPlaylistForm
+from apps.dashboard.forms import DashboardPlaylistForm
 from apps.dashboard.tables import PlaylistsListTable
 from apps.exercises.models import Playlist
 
@@ -24,40 +28,82 @@ def playlists_list_view(request):
 
 @login_required
 def playlist_add_view(request):
+    context = {
+        'verbose_name': Playlist._meta.verbose_name,
+        'verbose_name_plural': Playlist._meta.verbose_name_plural,
+    }
+
     if request.method == 'POST':
-        form = DashboardAddPlaylistForm(data=request.POST)
+        form = DashboardPlaylistForm(data=request.POST)
         form.context = {'user': request.user}
         if form.is_valid():
             playlist = form.save(commit=False)
             playlist.authored_by = request.user
             playlist.save()
-            return redirect('dashboard:playlists-list')
-    else:
-        form = DashboardAddPlaylistForm()
+            if 'save-and-continue' in request.POST:
+                success_url = reverse('dashboard:edit-playlist',
+                                      kwargs={'playlist_name': playlist.name})
+                messages.add_message(request, messages.SUCCESS,
+                                     f"{context['verbose_name']} has been saved successfully.")
+            else:
+                success_url = reverse('dashboard:playlists-list')
+            return redirect(success_url)
+        context['form'] = form
+        return render(request, "dashboard/content.html", context)
 
-    return render(request, "dashboard/playlist.html", {
-        "form": form
-    })
+    else:
+        form = DashboardPlaylistForm(initial=request.session.get('clone_data'))
+        request.session['clone_data'] = None
+
+    context['form'] = form
+    return render(request, "dashboard/content.html", context)
 
 
 @login_required
 def playlist_edit_view(request, playlist_name):
     playlist = get_object_or_404(Playlist, name=playlist_name)
 
+    context = {
+        'verbose_name': playlist._meta.verbose_name,
+        'verbose_name_plural': playlist._meta.verbose_name_plural,
+        'has_been_performed': playlist.has_been_performed,
+        'redirect_url': reverse('dashboard:playlists-list')
+    }
+
     if request.method == 'POST':
-        form = DashboardAddPlaylistForm(data=request.POST, instance=playlist)
+        form = DashboardPlaylistForm(data=request.POST, instance=playlist)
         form.context = {'user': request.user}
         if form.is_valid():
+            if 'save-as-new' in request.POST:
+                unique_fields = Playlist.get_unique_fields()
+                clone_data = copy(form.cleaned_data)
+                for field in clone_data:
+                    if field in unique_fields:
+                        clone_data[field] = None
+                request.session['clone_data'] = clone_data
+                return redirect('dashboard:add-playlist')
+
             playlist = form.save(commit=False)
             playlist.authored_by = request.user
             playlist.save()
-            return redirect('dashboard:playlists-list')
-    else:
-        form = DashboardAddPlaylistForm(instance=playlist)
+            if 'save-and-continue' in request.POST:
+                success_url = reverse('dashboard:edit-playlist',
+                                      kwargs={'playlist_name': playlist.name})
+                messages.add_message(request, messages.SUCCESS,
+                                     f"{context['verbose_name']} has been saved successfully.")
+            else:
+                success_url = reverse('dashboard:playlists-list')
+            return redirect(success_url)
+        context['form'] = form
+        return render(request, "dashboard/content.html", context)
 
-    return render(request, "dashboard/playlist.html", {
-        "form": form
-    })
+    if playlist.has_been_performed:
+        form = DashboardPlaylistForm(instance=playlist, disable_fields=True)
+    else:
+        form = DashboardPlaylistForm(instance=playlist)
+
+    context['form'] = form
+    return render(request, "dashboard/content.html", context)
 
 
 @login_required
@@ -66,8 +112,15 @@ def playlist_delete_view(request, playlist_name):
     if playlist.authored_by != request.user:
         raise PermissionDenied
 
-    if playlist.has_been_performed:
-        raise ValidationError('Playlists that have been performed cannot be deleted.')
+    if request.method == 'POST':
+        if playlist.has_been_performed:
+            raise ValidationError('Playlists that have been performed cannot be deleted.')
+        playlist.delete()
+        return redirect('dashboard:playlists-list')
 
-    playlist.delete()
-    return redirect('dashboard:playlists-list')
+    context = {'obj': playlist, 'obj_name': playlist.name,
+               'verbose_name': playlist._meta.verbose_name,
+               'verbose_name_plural': playlist._meta.verbose_name_plural,
+               'has_been_performed': playlist.has_been_performed,
+               'redirect_url': reverse('dashboard:playlists-list')}
+    return render(request, 'dashboard/delete-confirmation.html', context)
